fix: prevent panic when passing multiple directories and no files
When passing multiple directory only args, like `cw ./target/
/var/log/`, the program crashes with a panic due to trying to calculate
the log for zero (the `max_width` we got from finding the file with the
max amount of bytes). If all files are directories, the bytes are all 0.

```
cw: target/: Is a directory
thread 'main' panicked at src/main.rs:84:28:
argument of integer logarithm must be positive
note: run with `RUST_BACKTRACE=1` environment variable to display a backtrace
```

Solved by using the `total_bytes` instead and setting `max_width` to 1
if it's zero.