refactor(`cw:count`): split `count` function into `parse_content` and `parse_file`
`parse_content` takes a file name and the contents of the file then spits out a `File`. The file name is only used to construct a `File`, it doesn't actually check for the path. Might not be the best idea.
`parse_file` takes a file name, opens the file and runs the `parse_content` function on it. This function actually checks when it's a valid path.
This split was done so I could parse input from stdin.